pyproject.toml
[project]
name = "dsait-4065-conversational-agents"
version = "0.1.0"
description = "Add your description here"
authors = [
     { name = "Emiel", email = "emiel.witting@gmail.com" },
     { name = "Razo", email = "razovanberkel@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest",
    "deepface>=0.0.93",
    "mistralai>=1.5.1",
    "ollama>=0.4.7",
    "openai-whisper>=20240930",
    "pillow>=11.1.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
    "sounddevice>=0.5.1",
    "tf-keras>=2.19.0",
    "transformers>=4.49.0",
    "wavio>=0.0.9",
    "sv-ttk>=2.6.0",
    "gtts>=2.5.4",
    "playsound==1.2.2",
]
